+++ otj-metrics-actuator/src/test/java/com/opentable/metrics/actuate/AbstractTest.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.opentable.metrics.actuate;
+
+import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
+
+import org.junit.Before;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = TestMvcServer2Configuration.class, webEnvironment = RANDOM_PORT)
+@ActiveProfiles(profiles = "test")
+@TestPropertySource(properties = {
+    "spring.jmx.enabled=true",
+    "management.endpoints.web.exposure.include=*",
+    "management.endpoint.health.show-details=always",
+    "management.metrics.export.dw-new.enabled=true"
+})
+public abstract class AbstractTest {
+
+    @Autowired
+    protected WebApplicationContext context;
+
+    protected MockMvc mvc;
+
+    @Before
+    public void setUp() {
+        mvc = MockMvcBuilders.webAppContextSetup(context)
+            .build();
+    }
+
+}